flask/requirements.txt
flask==2.1.2
flask_cors==3.0.10
pillow==9.1.1